+++ tests/JwtAuthenticationTest.php
@@ -492,7 +492,7 @@ class JwtAuthenticationTest extends TestCase
         $collection = new MiddlewareCollection([
             new JwtAuthentication([
                 "secret" => "supersecretkeyyoushouldnotcommit",
-                "error" => function ($request, $response, $arguments) use (&$dummy) {
+                "error" => function (ResponseInterface $response, $arguments) use (&$dummy) {
                     $dummy = true;
                 }
             ])
@@ -521,7 +521,7 @@ class JwtAuthenticationTest extends TestCase
         $collection = new MiddlewareCollection([
             new JwtAuthentication([
                 "secret" => "supersecretkeyyoushouldnotcommittogithub",
-                "error" => function ($request, $response, $arguments) use (&$dummy) {
+                "error" => function (ResponseInterface $response, $arguments) use (&$dummy) {
                     $dummy = true;
                     $response->getBody()->write("Error");
                     return $response;