+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file, in reverse 
 - The `callback` setting was renamed to `before`. It is called before executing other middlewares in the stack.
 - The `passthrough` setting was renamed to `ignore`.
 - Public setter methods `addRule()` and `withRules()` are now immutable.
+- Error callback now receives only response and arguments, request was removed.
 
 ### Added
 - Support for the [approved version of PSR-15](https://github.com/php-fig/http-server-middleware).